- Updates against the newest Nebula dev branch. Some minor features had to be disabled, and few things removed from map as they are no longer present in Nebula.
- Nothing critical/serious should hopefully be affected.

+++ code/game/objects/items/weapons/tanks/tanks.dm
@@ -119,7 +119,7 @@ var/global/list/global/tank_gauge_cache = list()
 		icon = loc
 
 	if (istype(used_item, /obj/item/scanner/gas))
-		return TRUE
+		return FALSE // allow afterattack to proceed
 
 	if (istype(used_item,/obj/item/latexballon))
 		var/obj/item/latexballon/LB = used_item
@@ -177,9 +177,6 @@ var/global/list/global/tank_gauge_cache = list()
 		if(wired)
 			to_chat(user, "<span class='notice'>You begin attaching the assembly to \the [src].</span>")
 			if(do_after(user, 50, src))
-				to_chat(user, "<span class='notice'>You finish attaching the assembly to \the [src].</span>")
-				global.bombers += "[key_name(user)] attached an assembly to a wired [src]. Temp: [air_contents.temperature-T0C]"
-				log_and_message_admins("attached an assembly to a wired [src]. Temp: [air_contents.temperature-T0C]", user)
 				assemble_bomb(used_item,user)
 			else
 				to_chat(user, "<span class='notice'>You stop attaching the assembly.</span>")
@@ -512,7 +509,7 @@ var/global/list/global/tank_gauge_cache = list()
 	. = ..()
 
 	// Set up appearance/strings.
-	var/obj/item/tank/tank_copy = pick(typesof(/obj/item/tank/oxygen) + typesof(/obj/item/tank/hydrogen) + typesof(/obj/item/tank/phoron))
+	var/obj/item/tank/tank_copy = pick(typesof(/obj/item/tank/oxygen) + typesof(/obj/item/tank/hydrogen))
 	name = initial(tank_copy.name)
 	desc = initial(tank_copy.desc)
 	icon = initial(tank_copy.icon)
@@ -550,39 +547,45 @@ var/global/list/global/tank_gauge_cache = list()
 /obj/item/tankassemblyproxy/receive_signal()	//This is mainly called by the sensor through sense() to the holder, and from the holder to here.
 	tank.cause_explosion()	//boom (or not boom if you made shijwtty mix)
 
-/obj/item/tank/proc/assemble_bomb(used_item,user)	//Bomb assembly proc. This turns assembly+tank into a bomb
+/obj/item/tank/proc/assemble_bomb(used_item,mob/user)	//Bomb assembly proc. This turns assembly+tank into a bomb
 	var/obj/item/assembly_holder/S = used_item
-	var/mob/M = user
-	if(!S.secured)										//Check if the assembly is secured
-		return
 	if(isigniter(S.a_left) == isigniter(S.a_right))		//Check if either part of the assembly has an igniter, but if both parts are igniters, then fuck it
 		return
+	if(!S.secured)										//Check if the assembly is secured
+		to_chat(user, SPAN_NOTICE("\The [S] must be secured before attaching it to \the [src]!"))
+		return
 
-	if(!M.try_unequip(src))
+	if(!user.try_unequip(src))
 		return					//Remove the tank from your character,in case you were holding it
-	M.put_in_hands(src)			//Equips the bomb if possible, or puts it on the floor.
+	user.put_in_hands(src)			//Equips the bomb if possible, or puts it on the floor.
 
 	proxyassembly.assembly = S	//Tell the bomb about its assembly part
 	S.master = proxyassembly	//Tell the assembly about its new owner
-	S.forceMove(src)			//Move the assembly
+	user.remove_from_mob(S, src, FALSE) //Move the assembly and reset HUD layer/plane status
 
 	update_icon()
+	to_chat(user, "<span class='notice'>You finish attaching the assembly to \the [src].</span>")
+	global.bombers += "[key_name(user)] attached an assembly to a wired [src]. Temp: [air_contents.temperature-T0C]"
+	log_and_message_admins("attached an assembly to a wired [src]. Temp: [air_contents.temperature-T0C]", user)
 
 /obj/item/tank/proc/cause_explosion()	//This happens when a bomb is told to explode
+
 	var/obj/item/assembly_holder/assy = proxyassembly.assembly
-	var/ign = assy.a_right
-	var/obj/item/other = assy.a_left
+	var/obj/item/igniter = assy.a_right
+	var/obj/item/other   = assy.a_left
 
 	if (isigniter(assy.a_left))
-		ign = assy.a_left
-		other = assy.a_right
+		igniter = assy.a_left
+		other   = assy.a_right
 
 	if(other)
 		other.dropInto(get_turf(src))
-	qdel(ign)
+	if(!QDELETED(igniter))
+		qdel(igniter)
 	assy.master = null
 	proxyassembly.assembly = null
-	qdel(assy)
+	if(!QDELETED(assy))
+		qdel(assy)
 	update_icon()
 
 	air_contents.add_thermal_energy(15000)
